+++ Runtime/Attributes/LayerSelectorAttribute.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+
+namespace Mane.Unity
+{
+    public class LayerSelectorAttribute : PropertyAttribute { }
+}